requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
transformers>=4.30.0
matplotlib>=3.4.3
tqdm>=4.62.3 